Revert "chore: modularize workspace into subcrates (#66)"
This reverts commit e8435c549f7b6ce762e266de4f41d1813f505880.
