--- dead-man-email/Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-name = "dead-man-email"
-version = "0.1.0"
-edition = "2021"
-authors = ["Jose Storopoli <jose@storopoli.io>"]
-description = "Dead Man's Switch Email Library"
-license = "AGPL-3.0-only"
-
-[dependencies]
-dead-man-config = { path = "../dead-man-config", version = "0.1" }
-thiserror = "1"
-lettre = { version = "0.10", features = ["rustls-tls", "builder"] }
-lettre_email = "0.9"
-mime_guess = "2"
