--- dead-man-switch/Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-name = "dead-man-switch"
-version = "0.3.0"
-edition = "2021"
-authors = ["Jose Storopoli <jose@storopoli.io>"]
-description = "A simple no-BS Dead Man's Switch"
-license = "AGPL-3.0-only"
-readme = "../README.md"
-
-[dependencies]
-dead-man-config = { path = "../dead-man-config", version = "0.1" }
-dead-man-timers = { path = "../dead-man-timers", version = "0.1" }
-dead-man-email = { path = "../dead-man-email", version = "0.1" }
-dead-man-tui = { path = "../dead-man-tui", version = "0.1" }
