--- dead-man-timers/Cargo.toml
@@ -1,11 +1,0 @@
-[package]
-name = "dead-man-timers"
-version = "0.1.0"
-edition = "2021"
-authors = ["Jose Storopoli <jose@storopoli.io>"]
-description = "Dead Men's Switch Timer Library"
-license = "AGPL-3.0-only"
-
-[dependencies]
-dead-man-config = { path = "../dead-man-config", version = "0.1" }
-chrono = "0.4"
