+++ src/main.rs
@@ -8,12 +8,14 @@
 
 use std::error::Error;
 
-use dead_man_tui::run;
+#[cfg(feature = "tui")]
+use dead_man_switch::run;
 
 /// The main function.
 ///
 /// This function executes the main loop of the application
 /// by calling the [`run`] function.
+#[cfg(feature = "tui")]
 fn main() -> Result<(), Box<dyn Error>> {
     run()?;
     Ok(())
